Moved all headers down a folder. Begun namespace changes and renaming.
