+++ include/NovelRT/Graphics/NovelBasicFillRect.h
@@ -1,15 +1,15 @@
 // Copyright © Matt Jones and Contributors. Licensed under the MIT License (MIT). See LICENCE.md in the repository root for more information.
 
-#ifndef NOVELRT_NOVELRECT_H
-#define NOVELRT_NOVELRECT_H
+#ifndef NOVELRT_GRAPHICS_NOVELRECT_H
+#define NOVELRT_GRAPHICS_NOVELRECT_H
 
 #include "NovelRenderObject.h"
 #include "RGBAConfig.h"
 #include "ShaderProgram.h"
 #include "NovelCamera.h"
 
-namespace NovelRT {
-  class NovelBasicFillRect : public NovelRenderObject {
+namespace NovelRT::Graphics {
+  class BasicFillRect : public NovelRenderObject {
 
   private:
     RGBAConfig _colourConfig;
@@ -20,7 +20,7 @@ namespace NovelRT {
     void configureObjectBuffers() final;
 
   public:
-    NovelBasicFillRect(NovelLayeringService* layeringService,
+    BasicFillRect(NovelLayeringService* layeringService,
                        const RGBAConfig& fillColour,
                        const NovelCommonArgs& args,
                        ShaderProgram shaderProgram,
@@ -32,4 +32,4 @@ namespace NovelRT {
   };
 }
 
-#endif //NOVELRT_NOVELRECT_H
+#endif //NOVELRT_GRAPHICS_NOVELRECT_H
