+++ include/NovelRT/Graphics/NovelImageRect.h
