+++ include/NovelRT/Graphics/NovelRenderObject.h
