+++ include/NovelRT/Graphics/NovelTextRect.h
