+++ include/NovelRT/Graphics/RGBAConfig.h
