+++ include/NovelRT/Input/NovelBasicInteractionRect.h
