+++ include/NovelRT/Input/NovelInteractionObject.h
