+++ include/NovelRT/Maths/GeoBounds.h
