+++ include/NovelRT/NovelLayeringService.h
