+++ include/NovelRT/NovelLoggingService.h
