+++ include/NovelRT/NovelWorldObject.h
