+++ include/NovelRT/NovelWorldObjectSortComparison.h
