+++ include/NovelRT/Utilities/Lazy.h
