+++ include/NovelRT/Utilities/NovelCommonArgs.h
