+++ include/NovelRT/Utilities/NovelUtilities.h
