+++ include/NovelRT/Windowing/NovelWindowingService.h
